Separate directory for directives
