+++ client/src/app/commons/directives/unique-email/unique-email.directive.spec.js
